FARS Accident data set
The data set had been previously cleaned and prepared for analysis by the project team according to the purpose of this project. The cleaned version contains about 28,000 observations and 20 variables. Therefore it is not equal to the original FARS Accident data set for 2015, which contains more than 60 variables and abour 30,000 observations.